+++ backend/canisters/multi_user/impl/src/guards.rs
@@ -25,3 +25,11 @@ pub fn caller_is_local_user_index() -> Result<(), String> {
         Err("Caller is not the local_user_index canister".to_owned())
     }
 }
+
+pub fn caller_is_user_index() -> Result<(), String> {
+    if read_state(|state| state.is_caller_user_index()) {
+        Ok(())
+    } else {
+        Err("Caller is not the user_index canister".to_owned())
+    }
+}