+++ backend/canisters/multi_user/impl/src/lib.rs
@@ -49,6 +49,10 @@ impl RuntimeState {
         self.env.caller() == self.data.local_user_index_canister_id
     }
 
+    pub fn is_caller_user_index(&self) -> bool {
+        self.env.caller() == self.data.user_index_canister_id
+    }
+
     // The index of the user the caller owns, if the caller is one of this canister's users
     pub fn caller_user_index(&self) -> Option<u16> {
         self.data.users.index_by_principal(&self.env.caller())